Refactor tests orgs
