+++ common/dal2/src/lib.rs
@@ -20,3 +20,10 @@ pub mod error;
 pub mod services;
 
 pub use dal::DataAccessor;
+
+#[cfg(test)]
+mod tests {
+    mod ops {
+        mod io;
+    }
+}
